backend-vercel/requirements.txt
fastapi==0.115.12
google-genai==1.34.0
